// src/types/resource.type.ts
export type ResourceType = "people" | "films" | "starships" | "planets";

export type Resource =
  | StarWarsCharacter
  | StarWarsFilm
  | StarWarsStarship
  | StarWarsPlanet;

export interface StarWarsResource {
  id: string;
  url: string;
  created: string;
  edited: string;
  [key: string]: unknown;
}

export interface StarWarsCharacter extends StarWarsResource {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
}

export interface StarWarsFilm extends StarWarsResource {
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  characters: string[];
  planets: string[];
  starships: string[];
  vehicles: string[];
  species: string[];
}

export interface StarWarsStarship extends StarWarsResource {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  consumables: string;
  hyperdrive_rating: string;
  MGLT: string;
  starship_class: string;
  pilots: string[];
  films: string[];
}

export interface StarWarsPlanet extends StarWarsResource {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
}